[Fix] Fix EncodeIEEE754binary64.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@4390 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ common/Endianness.h
@@ -192,7 +192,7 @@ static forceinline uint64 EncodeIEEE754binary64(float64 f)
 		return conv.i;
 	#else
 		uint64 i = 0;
-		std::memcpy(&i, &f, sizeof(float32));
+		std::memcpy(&i, &f, sizeof(float64));
 		return i;
 	#endif
 #else
